# app/Models/Artist.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Factories\HasFactory;

/**
 * @OA\Schema(
 *     schema="Artist",
 *     type="object",
 *     title="Artist",
 *     description="Schema for Artist",
 *     required={"id", "name"},
 *     @OA\Property(
 *         property="id",
 *         type="integer",
 *         description="Unique identifier for the artist",
 *         example=1
 *     ),
 *     @OA\Property(
 *         property="name",
 *         type="string",
 *         description="Name of the artist",
 *         example="The Beatles"
 *     ),
 *     @OA\Property(
 *         property="created_at",
 *         type="string",
 *         format="date-time",
 *         description="Date when the artist was created",
 *         example="2024-11-15T12:00:00Z"
 *     ),
 *     @OA\Property(
 *         property="updated_at",
 *         type="string",
 *         format="date-time",
 *         description="Date when the artist was last updated",
 *         example="2024-11-15T12:00:00Z"
 *     )
 * )
 */

class Artist extends Model
{
    use HasFactory;

    protected $fillable = ['name'];

    public function albums()
    {
        return $this->hasMany(Album::class);
    }
}